+++ src/Servers/Kestrel/Core/src/Internal/Http/Http1Connection.cs
@@ -44,8 +44,6 @@ namespace Microsoft.AspNetCore.Server.Kestrel.Core.Internal.Http
             _keepAliveTicks = ServerOptions.Limits.KeepAliveTimeout.Ticks;
             _requestHeadersTimeoutTicks = ServerOptions.Limits.RequestHeadersTimeout.Ticks;
 
-            RequestBodyPipe = CreateRequestBodyPipe();
-
             _http1Output = new Http1OutputProducer(
                 _context.Transport.Output,
                 _context.ConnectionId,
@@ -57,6 +55,7 @@ namespace Microsoft.AspNetCore.Server.Kestrel.Core.Internal.Http
 
             Input = _context.Transport.Input;
             Output = _http1Output;
+            MemoryPool = _context.MemoryPool;
         }
 
         public PipeReader Input { get; }
@@ -66,6 +65,8 @@ namespace Microsoft.AspNetCore.Server.Kestrel.Core.Internal.Http
         public MinDataRate MinRequestBodyDataRate { get; set; }
 
         public MinDataRate MinResponseDataRate { get; set; }
+
+        public MemoryPool<byte> MemoryPool { get; }
 
         protected override void OnRequestProcessingEnded()
         {
@@ -531,17 +532,5 @@ namespace Microsoft.AspNetCore.Server.Kestrel.Core.Internal.Http
         }
 
         void IRequestProcessor.Tick(DateTimeOffset now) { }
-
-        private Pipe CreateRequestBodyPipe()
-            => new Pipe(new PipeOptions
-            (
-                pool: _context.MemoryPool,
-                readerScheduler: ServiceContext.Scheduler,
-                writerScheduler: PipeScheduler.Inline,
-                pauseWriterThreshold: 1,
-                resumeWriterThreshold: 1,
-                useSynchronizationContext: false,
-                minimumSegmentSize: KestrelMemoryPool.MinimumSegmentSize
-            ));
     }
 }
