+++ src/Servers/Kestrel/Core/src/Internal/Http/Http1MessageBody.cs
@@ -2,21 +2,16 @@
 // Licensed under the Apache License, Version 2.0. See License.txt in the project root for license information.
 
 using System;
-using System.Buffers;
-using System.Diagnostics;
-using System.IO;
 using System.IO.Pipelines;
-using System.Threading;
 using System.Threading.Tasks;
 using Microsoft.AspNetCore.Connections;
-using Microsoft.AspNetCore.Http;
 using Microsoft.AspNetCore.Server.Kestrel.Core.Internal.Infrastructure;
 
 namespace Microsoft.AspNetCore.Server.Kestrel.Core.Internal.Http
 {
     public abstract class Http1MessageBody : MessageBody
     {
-        private readonly Http1Connection _context;
+        protected readonly Http1Connection _context;
 
         protected Http1MessageBody(Http1Connection context)
             : base(context, context.MinRequestBodyDataRate)
@@ -83,29 +78,6 @@ namespace Microsoft.AspNetCore.Server.Kestrel.Core.Internal.Http
             }
         }
 
-        protected void Copy(ReadOnlySequence<byte> readableBuffer, PipeWriter writableBuffer)
-        {
-            if (readableBuffer.IsSingleSegment)
-            {
-                writableBuffer.Write(readableBuffer.First.Span);
-            }
-            else
-            {
-                foreach (var memory in readableBuffer)
-                {
-                    writableBuffer.Write(memory.Span);
-                }
-            }
-        }
-
-        public abstract void Advance(long consumedBytes);
-
-
-        protected virtual bool Read(ReadOnlySequence<byte> readableBuffer, PipeWriter writableBuffer, out SequencePosition consumed, out SequencePosition examined)
-        {
-            throw new NotImplementedException();
-        }
-
         public static MessageBody For(
             HttpVersion httpVersion,
             HttpRequestHeaders headers,
@@ -151,6 +123,8 @@ namespace Microsoft.AspNetCore.Server.Kestrel.Core.Internal.Http
                 }
 
                 context.RequestBodyPipeReader = new HttpRequestPipeReader();
+                // TODO may push more into the wrapper rather than just calling into the message body
+                // NBD for now.
                 return new ForChunkedEncoding(keepAlive, context);
             }
 
@@ -176,429 +150,6 @@ namespace Microsoft.AspNetCore.Server.Kestrel.Core.Internal.Http
             }
 
             return keepAlive ? MessageBody.ZeroContentLengthKeepAlive : MessageBody.ZeroContentLengthClose;
-        }
-
-        /// <summary>
-        /// The upgrade stream uses the raw connection stream instead of going through the RequestBodyPipe. This
-        /// removes the redundant copy from the transport pipe to the body pipe.
-        /// </summary>
-        private class ForUpgrade : Http1MessageBody
-        {
-            public ForUpgrade(Http1Connection context)
-                : base(context)
-            {
-                RequestUpgrade = true;
-            }
-
-            // This returns IsEmpty so we can avoid draining the body (since it's basically an endless stream)
-            public override bool IsEmpty => true;
-
-            public override ValueTask<ReadResult> ReadAsync(CancellationToken cancellationToken = default)
-            {
-                return _context.Input.ReadAsync(cancellationToken);
-            }
-
-            public override bool TryRead(out ReadResult result)
-            {
-                return _context.Input.TryRead(out result);
-            }
-
-            public override void AdvanceTo(SequencePosition consumed)
-            {
-                _context.Input.AdvanceTo(consumed);
-            }
-
-            public override void AdvanceTo(SequencePosition consumed, SequencePosition examined)
-            {
-                _context.Input.AdvanceTo(consumed, examined);
-            }
-
-            public override void Advance(long consumedBytes)
-            {
-            }
-
-            public override void Complete(Exception exception)
-            {
-                // Noop as we don't want to complete the connection pipe.
-            }
-
-            public override void CancelPendingRead()
-            {
-                _context.Input.CancelPendingRead();
-            }
-
-            public override void OnWriterCompleted(Action<Exception, object> callback, object state)
-            {
-                _context.Input.OnWriterCompleted(callback, state);
-            }
-
-            public override Task ConsumeAsync()
-            {
-                return Task.CompletedTask;
-            }
-
-            public override Task StopAsync()
-            {
-                return Task.CompletedTask;
-            }
-        }
-
-        private class ForContentLength : Http1MessageBody
-        {
-            private readonly long _contentLength;
-            private long _inputLength;
-
-            public ForContentLength(bool keepAlive, long contentLength, Http1Connection context)
-                : base(context)
-            {
-                RequestKeepAlive = keepAlive;
-                _contentLength = contentLength;
-                _inputLength = _contentLength;
-            }
-
-            protected override bool Read(ReadOnlySequence<byte> readableBuffer, PipeWriter writableBuffer, out SequencePosition consumed, out SequencePosition examined)
-            {
-                if (_inputLength == 0)
-                {
-                    throw new InvalidOperationException("Attempted to read from completed Content-Length request body.");
-                }
-
-                var actual = (int)Math.Min(readableBuffer.Length, _inputLength);
-                _inputLength -= actual;
-
-                consumed = readableBuffer.GetPosition(actual);
-                examined = consumed;
-
-                Copy(readableBuffer.Slice(0, actual), writableBuffer);
-
-                return _inputLength == 0;
-            }
-
-            protected override void OnReadStarting()
-            {
-                if (_contentLength > _context.MaxRequestBodySize)
-                {
-                    BadHttpRequestException.Throw(RequestRejectionReason.RequestBodyTooLarge);
-                }
-            }
-        }
-
-        /// <summary>
-        ///   http://tools.ietf.org/html/rfc2616#section-3.6.1
-        /// </summary>
-        private class ForChunkedEncoding : Http1MessageBody
-        {
-            // byte consts don't have a data type annotation so we pre-cast it
-            private const byte ByteCR = (byte)'\r';
-            // "7FFFFFFF\r\n" is the largest chunk size that could be returned as an int.
-            private const int MaxChunkPrefixBytes = 10;
-
-            private long _inputLength;
-
-            private Mode _mode = Mode.Prefix;
-
-            public ForChunkedEncoding(bool keepAlive, Http1Connection context)
-                : base(context)
-            {
-                RequestKeepAlive = keepAlive;
-            }
-
-            protected override bool Read(ReadOnlySequence<byte> readableBuffer, PipeWriter writableBuffer, out SequencePosition consumed, out SequencePosition examined)
-            {
-                consumed = default(SequencePosition);
-                examined = default(SequencePosition);
-
-                while (_mode < Mode.Trailer)
-                {
-                    if (_mode == Mode.Prefix)
-                    {
-                        ParseChunkedPrefix(readableBuffer, out consumed, out examined);
-
-                        if (_mode == Mode.Prefix)
-                        {
-                            return false;
-                        }
-
-                        readableBuffer = readableBuffer.Slice(consumed);
-                    }
-
-                    if (_mode == Mode.Extension)
-                    {
-                        ParseExtension(readableBuffer, out consumed, out examined);
-
-                        if (_mode == Mode.Extension)
-                        {
-                            return false;
-                        }
-
-                        readableBuffer = readableBuffer.Slice(consumed);
-                    }
-
-                    if (_mode == Mode.Data)
-                    {
-                        ReadChunkedData(readableBuffer, writableBuffer, out consumed, out examined);
-
-                        if (_mode == Mode.Data)
-                        {
-                            return false;
-                        }
-
-                        readableBuffer = readableBuffer.Slice(consumed);
-                    }
-
-                    if (_mode == Mode.Suffix)
-                    {
-                        ParseChunkedSuffix(readableBuffer, out consumed, out examined);
-
-                        if (_mode == Mode.Suffix)
-                        {
-                            return false;
-                        }
-
-                        readableBuffer = readableBuffer.Slice(consumed);
-                    }
-                }
-
-                // Chunks finished, parse trailers
-                if (_mode == Mode.Trailer)
-                {
-                    ParseChunkedTrailer(readableBuffer, out consumed, out examined);
-
-                    if (_mode == Mode.Trailer)
-                    {
-                        return false;
-                    }
-
-                    readableBuffer = readableBuffer.Slice(consumed);
-                }
-
-                // _consumedBytes aren't tracked for trailer headers, since headers have separate limits.
-                if (_mode == Mode.TrailerHeaders)
-                {
-                    if (_context.TakeMessageHeaders(readableBuffer, out consumed, out examined))
-                    {
-                        _mode = Mode.Complete;
-                    }
-                }
-
-                return _mode == Mode.Complete;
-            }
-
-            private void ParseChunkedPrefix(ReadOnlySequence<byte> buffer, out SequencePosition consumed, out SequencePosition examined)
-            {
-                consumed = buffer.Start;
-                examined = buffer.Start;
-                var reader = new SequenceReader<byte>(buffer);
-
-                if (!reader.TryRead(out var ch1) || !reader.TryRead(out var ch2))
-                {
-                    examined = reader.Position;
-                    return;
-                }
-
-                var chunkSize = CalculateChunkSize(ch1, 0);
-                ch1 = ch2;
-
-                while (reader.Consumed < MaxChunkPrefixBytes)
-                {
-                    if (ch1 == ';')
-                    {
-                        consumed = reader.Position;
-                        examined = reader.Position;
-
-                        AddAndCheckConsumedBytes(reader.Consumed);
-                        _inputLength = chunkSize;
-                        _mode = Mode.Extension;
-                        return;
-                    }
-
-                    if (!reader.TryRead(out ch2))
-                    {
-                        examined = reader.Position;
-                        return;
-                    }
-
-                    if (ch1 == '\r' && ch2 == '\n')
-                    {
-                        consumed = reader.Position;
-                        examined = reader.Position;
-
-                        AddAndCheckConsumedBytes(reader.Consumed);
-                        _inputLength = chunkSize;
-                        _mode = chunkSize > 0 ? Mode.Data : Mode.Trailer;
-                        return;
-                    }
-
-                    chunkSize = CalculateChunkSize(ch1, chunkSize);
-                    ch1 = ch2;
-                }
-
-                // At this point, 10 bytes have been consumed which is enough to parse the max value "7FFFFFFF\r\n".
-                BadHttpRequestException.Throw(RequestRejectionReason.BadChunkSizeData);
-            }
-
-            private void ParseExtension(ReadOnlySequence<byte> buffer, out SequencePosition consumed, out SequencePosition examined)
-            {
-                // Chunk-extensions not currently parsed
-                // Just drain the data
-                consumed = buffer.Start;
-                examined = buffer.Start;
-
-                do
-                {
-                    SequencePosition? extensionCursorPosition = buffer.PositionOf(ByteCR);
-                    if (extensionCursorPosition == null)
-                    {
-                        // End marker not found yet
-                        consumed = buffer.End;
-                        examined = buffer.End;
-                        AddAndCheckConsumedBytes(buffer.Length);
-                        return;
-                    };
-
-                    var extensionCursor = extensionCursorPosition.Value;
-                    var charsToByteCRExclusive = buffer.Slice(0, extensionCursor).Length;
-
-                    var suffixBuffer = buffer.Slice(extensionCursor);
-                    if (suffixBuffer.Length < 2)
-                    {
-                        consumed = extensionCursor;
-                        examined = buffer.End;
-                        AddAndCheckConsumedBytes(charsToByteCRExclusive);
-                        return;
-                    }
-
-                    suffixBuffer = suffixBuffer.Slice(0, 2);
-                    var suffixSpan = suffixBuffer.ToSpan();
-
-                    if (suffixSpan[1] == '\n')
-                    {
-                        // We consumed the \r\n at the end of the extension, so switch modes.
-                        _mode = _inputLength > 0 ? Mode.Data : Mode.Trailer;
-
-                        consumed = suffixBuffer.End;
-                        examined = suffixBuffer.End;
-                        AddAndCheckConsumedBytes(charsToByteCRExclusive + 2);
-                    }
-                    else
-                    {
-                        // Don't consume suffixSpan[1] in case it is also a \r.
-                        buffer = buffer.Slice(charsToByteCRExclusive + 1);
-                        consumed = extensionCursor;
-                        AddAndCheckConsumedBytes(charsToByteCRExclusive + 1);
-                    }
-                } while (_mode == Mode.Extension);
-            }
-
-            private void ReadChunkedData(ReadOnlySequence<byte> buffer, PipeWriter writableBuffer, out SequencePosition consumed, out SequencePosition examined)
-            {
-                var actual = Math.Min(buffer.Length, _inputLength);
-                consumed = buffer.GetPosition(actual);
-                examined = consumed;
-
-                Copy(buffer.Slice(0, actual), writableBuffer);
-
-                _inputLength -= actual;
-                AddAndCheckConsumedBytes(actual);
-
-                if (_inputLength == 0)
-                {
-                    _mode = Mode.Suffix;
-                }
-            }
-
-            private void ParseChunkedSuffix(ReadOnlySequence<byte> buffer, out SequencePosition consumed, out SequencePosition examined)
-            {
-                consumed = buffer.Start;
-                examined = buffer.Start;
-
-                if (buffer.Length < 2)
-                {
-                    examined = buffer.End;
-                    return;
-                }
-
-                var suffixBuffer = buffer.Slice(0, 2);
-                var suffixSpan = suffixBuffer.ToSpan();
-                if (suffixSpan[0] == '\r' && suffixSpan[1] == '\n')
-                {
-                    consumed = suffixBuffer.End;
-                    examined = suffixBuffer.End;
-                    AddAndCheckConsumedBytes(2);
-                    _mode = Mode.Prefix;
-                }
-                else
-                {
-                    BadHttpRequestException.Throw(RequestRejectionReason.BadChunkSuffix);
-                }
-            }
-
-            private void ParseChunkedTrailer(ReadOnlySequence<byte> buffer, out SequencePosition consumed, out SequencePosition examined)
-            {
-                consumed = buffer.Start;
-                examined = buffer.Start;
-
-                if (buffer.Length < 2)
-                {
-                    examined = buffer.End;
-                    return;
-                }
-
-                var trailerBuffer = buffer.Slice(0, 2);
-                var trailerSpan = trailerBuffer.ToSpan();
-
-                if (trailerSpan[0] == '\r' && trailerSpan[1] == '\n')
-                {
-                    consumed = trailerBuffer.End;
-                    examined = trailerBuffer.End;
-                    AddAndCheckConsumedBytes(2);
-                    _mode = Mode.Complete;
-                }
-                else
-                {
-                    _mode = Mode.TrailerHeaders;
-                }
-            }
-
-            private int CalculateChunkSize(int extraHexDigit, int currentParsedSize)
-            {
-                try
-                {
-                    checked
-                    {
-                        if (extraHexDigit >= '0' && extraHexDigit <= '9')
-                        {
-                            return currentParsedSize * 0x10 + (extraHexDigit - '0');
-                        }
-                        else if (extraHexDigit >= 'A' && extraHexDigit <= 'F')
-                        {
-                            return currentParsedSize * 0x10 + (extraHexDigit - ('A' - 10));
-                        }
-                        else if (extraHexDigit >= 'a' && extraHexDigit <= 'f')
-                        {
-                            return currentParsedSize * 0x10 + (extraHexDigit - ('a' - 10));
-                        }
-                    }
-                }
-                catch (OverflowException ex)
-                {
-                    throw new IOException(CoreStrings.BadRequest_BadChunkSizeData, ex);
-                }
-
-                BadHttpRequestException.Throw(RequestRejectionReason.BadChunkSizeData);
-                return -1; // can't happen, but compiler complains
-            }
-
-            private enum Mode
-            {
-                Prefix,
-                Extension,
-                Data,
-                Suffix,
-                Trailer,
-                TrailerHeaders,
-                Complete
-            };
         }
     }
 }
