+++ src/Servers/Kestrel/Core/src/Internal/Http2/Http2MessageBody.cs
@@ -1,15 +1,18 @@
 // Copyright (c) .NET Foundation. All rights reserved.
 // Licensed under the Apache License, Version 2.0. See License.txt in the project root for license information.
 
+using System;
+using System.IO.Pipelines;
+using System.Threading;
 using System.Threading.Tasks;
 using Microsoft.AspNetCore.Server.Kestrel.Core.Internal.Http;
-using Microsoft.AspNetCore.Server.Kestrel.Core.Internal.Infrastructure;
 
 namespace Microsoft.AspNetCore.Server.Kestrel.Core.Internal.Http2
 {
     public class Http2MessageBody : MessageBody
     {
         private readonly Http2Stream _context;
+        private ReadResult _previousReadResult;
 
         private Http2MessageBody(Http2Stream context, MinDataRate minRequestBodyDataRate)
             : base(context, minRequestBodyDataRate)
@@ -50,6 +53,61 @@ namespace Microsoft.AspNetCore.Server.Kestrel.Core.Internal.Http2
             }
 
             return new Http2MessageBody(context, minRequestBodyDataRate);
+        }
+
+        public override void AdvanceTo(SequencePosition consumed)
+        {
+            AdvanceTo(consumed, consumed);
+        }
+
+        public override void AdvanceTo(SequencePosition consumed, SequencePosition examined)
+        {
+            var dataLength = _previousReadResult.Buffer.Slice(_previousReadResult.Buffer.Start, consumed).Length;
+            _context.RequestBodyPipe.Reader.AdvanceTo(consumed, examined);
+            OnDataRead(dataLength);
+        }
+
+        public override bool TryRead(out ReadResult readResult)
+        {
+            return _context.RequestBodyPipe.Reader.TryRead(out readResult);
+        }
+
+        public override async ValueTask<ReadResult> ReadAsync(CancellationToken cancellationToken = default)
+        {
+            _previousReadResult = await StartTimingReadAsync(cancellationToken);
+            StopTimingRead(_previousReadResult.Buffer.Length);
+
+            if (_previousReadResult.IsCompleted)
+            {
+                TryStop();
+            }
+
+            return _previousReadResult;
+        }
+
+        private ValueTask<ReadResult> StartTimingReadAsync(CancellationToken cancellationToken)
+        {
+            var readAwaitable = _context.RequestBodyPipe.Reader.ReadAsync(cancellationToken);
+
+            if (!readAwaitable.IsCompleted && _timingEnabled)
+            {
+                _backpressure = true;
+                _context.TimeoutControl.StartTimingRead();
+            }
+
+            return readAwaitable;
+        }
+
+        private void StopTimingRead(long bytesRead)
+        {
+            _context.TimeoutControl.BytesRead(bytesRead - _alreadyTimedBytes);
+            _alreadyTimedBytes = 0;
+
+            if (_backpressure)
+            {
+                _backpressure = false;
+                _context.TimeoutControl.StopTimingRead();
+            }
         }
     }
 }
